* merged r805, r805, r810, r811, r812, r813, r814, r815, r816, r817, r825, 	  r827, r847, r851, r868, r868, r875, and r883 from trunk:           * fixed hardcoded lcms to use LCMSLIBS, as pointed out by Felix von Leitner           * higher default angle accurancy           * sacrified a little-performance while encoding TIFFs by inverting             1-bit data to MINISWHIE to please some historic junk applications,             including Apple's Preview.app included in OS X 10.4           * optimized inverting 1bit and 8bit image data           * fixed serveral aspects of the new on-the-fly invertion code of             the TIFCodec, including invalid free, and format conditionals, oh my           * fixed saving non single bit images as MINISBLACK, regression             from the former commit that was intended to save single bit             data as MINISWHITE, inverting the data on-the-fly           * invert all MINISWHITE files while decoding, not just 1bit data           * fixed (very tiny) memory leaks in the JPEG codec spotted by valgrind             (while I was looking for something else), probably only slipped in             due to the endless rewrites of the partial JPEG decoding and             transforming that is so ugly due to libjpeg uglyness, ...           * fixed codecs to not segfault on read/write without any codec             registered           * fixed the TIFF codec to actually really use G4 compression if             G4 compression is requested           * get rid of the switch warnings, for now           * made DataMatrix dtor virtual           * fixed edentify physical size computation, ieek!           * fixed TIFCodec to zero-initialize the resolution if the TIFF             stream contains none           * fixed copy_crop_rotate to also allow negative start coordinates           * get rid of unused dnd x/y warnings           * moved the edisplay dnd x/y variable initialization to the correct location

+++ lib/rotate.cc
@@ -370,7 +370,7 @@ void rotate (Image& image, double angle, Image::iterator background)
   image.setRawData ();
 }
 
-Image* copy_crop_rotate (Image& image, unsigned int x_start, unsigned int y_start,
+Image* copy_crop_rotate (Image& image, int x_start, int y_start,
 			 unsigned int w, unsigned int h,
 			 double angle)
 {
@@ -392,13 +392,13 @@ Image* copy_crop_rotate (Image& image, unsigned int x_start, unsigned int y_star
   const double cached_sin = sin (angle);
   const double cached_cos = cos (angle);
 
-  Image::iterator background = image.begin();  background.setL (0);
+  Image::iterator background = image.begin(); background.setL (0);
   
-  for(int y = 0; y < h; ++y)
-      for(int x = 0; x < w; ++x)
+  for (int y = 0; y < h; ++y)
+      for (int x = 0; x < w; ++x)
 	{
-	  const double ox = (  x * cached_cos + y * cached_sin) + x_start;
-	  const double oy = (- x * cached_sin + y * cached_cos) + y_start;
+	  const double ox = ((double) x * cached_cos + y * cached_sin) + x_start;
+	  const double oy = ((double)-x * cached_sin + y * cached_cos) + y_start;
 	  
 	  if (ox >= 0 && oy >= 0 &&
 	      ox < image.w && oy < image.h) {
